ChaCuN: It's not a bug...
After only 276 commits, 500 immutability problems, and a personal relationship with the JavaFX docs, we have finally reached the end of this project. The Bugs have won, features are no more, and our sanity has been stretched to its limits. Celebrate the culmination of countless "just one more thing" moments and the birth of a codebase that should probably come with a warning label.